+++ src/Optimisarr.Data/Migrations/20260628140959_AddDolbyVisionHandling.cs
@@ -1,0 +1,40 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace Optimisarr.Data.Migrations
+{
+    /// <inheritdoc />
+    public partial class AddDolbyVisionHandling : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<bool>(
+                name: "IsDolbyVision",
+                table: "MediaFiles",
+                type: "INTEGER",
+                nullable: false,
+                defaultValue: false);
+
+            migrationBuilder.AddColumn<bool>(
+                name: "OptimiseDolbyVision",
+                table: "Libraries",
+                type: "INTEGER",
+                nullable: false,
+                defaultValue: false);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "IsDolbyVision",
+                table: "MediaFiles");
+
+            migrationBuilder.DropColumn(
+                name: "OptimiseDolbyVision",
+                table: "Libraries");
+        }
+    }
+}